requirements.txt
tk==0.1.0
pymysql==1.0.2
pillow==9.4.0